feat: Separating http libraries
