+++ src/Uno.Extensions-packageonly.slnf
@@ -24,7 +24,8 @@
       "Uno.Extensions.Hosting.UI\\Uno.Extensions.Hosting.WinUI.csproj",
       "Uno.Extensions.Hosting\\Uno.Extensions.Hosting.csproj",
       "Uno.Extensions.Http.Refit\\Uno.Extensions.Http.Refit.csproj",
-      "Uno.Extensions.Http.Tests\\Uno.Extensions.Http.Tests.csproj",
+      "Uno.Extensions.Http.UI\\Uno.Extensions.Http.UI.csproj",
+      "Uno.Extensions.Http.UI\\Uno.Extensions.Http.WinUI.csproj",
       "Uno.Extensions.Http\\Uno.Extensions.Http.csproj",
       "Uno.Extensions.Localization.UI\\Uno.Extensions.Localization.UI.csproj",
       "Uno.Extensions.Localization.UI\\Uno.Extensions.Localization.WinUI.csproj",
