+++ src/Uno.Extensions.Http.Refit/ServiceCollectionExtensions.cs
@@ -1,13 +1,25 @@
-﻿using System;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Refit;
 using Uno.Extensions.Http;
 
 namespace Uno.Extensions;
 
+/// <summary>
+/// Extensions for <see cref="IServiceCollection"/>
+/// </summary>
 public static class ServiceCollectionExtensions
 {
+	/// <summary>
+	/// Registers a Refit client with the specified <paramref name="name"/>.
+	/// </summary>
+	/// <typeparam name="TInterface">The Refit api type to register</typeparam>
+	/// <param name="services">The services collection to register the api with</param>
+	/// <param name="context">The host builder context</param>
+	/// <param name="name">[optional]The name for locating endpoint information in appsettings</param>
+	/// <param name="settingsBuilder">[optional]Callback for overriding Refit settings</param>
+	/// <param name="configure">[optional]Callback for configuring the endpoint</param>
+	/// <returns></returns>
 	public static IServiceCollection AddRefitClient<TInterface>(
 		   this IServiceCollection services,
 		   HostBuilderContext context,
