+++ src/Uno.Extensions.Http.UI/HostBuilderExtensions.cs
@@ -1,14 +1,29 @@
 ﻿namespace Uno.Extensions;
 
+/// <summary>
+/// Extensions for <see cref="IHostBuilder"/>
+/// </summary>
 public static class HostBuilderExtensions
 {
+	/// <summary>
+	/// Registers the native HttpMessageHandler and the DiagnosticHandler
+	/// </summary>
+	/// <param name="builder">The host builder to register with</param>
+	/// <param name="configure">Callback for configuring Http services</param>
+	/// <returns>Updated host builder</returns>
 	public static IHostBuilder UseHttp(
-		this IHostBuilder hostBuilder,
+		this IHostBuilder builder,
 		Action<IServiceCollection> configure)
 	{
-		return hostBuilder.UseHttp((context, builder) => configure.Invoke(builder));
+		return builder.UseHttp((context, builder) => configure.Invoke(builder));
 	}
 
+	/// <summary>
+	/// Registers the native HttpMessageHandler and the DiagnosticHandler
+	/// </summary>
+	/// <param name="builder">The host builder to register with</param>
+	/// <param name="configure">Callback for configuring Http services</param>
+	/// <returns>Updated host builder</returns>
 	public static IHostBuilder UseHttp(
 		this IHostBuilder builder,
 		Action<HostBuilderContext, IServiceCollection>? configure = default)
@@ -19,6 +34,7 @@ public static class HostBuilderExtensions
 			if (!ctx.IsRegistered(nameof(UseHttp)))
 			{
 				_ = services
+					.AddCookieManager(ctx)
 					.AddNativeHandler(ctx)
 					.AddTransient<DelegatingHandler, DiagnosticHandler>();
 			}
