+++ src/Uno.Extensions.Http/GlobalUsings.cs
@@ -1,6 +1,7 @@
 ﻿global using System;
 global using System.IO;
 global using System.Linq;
+global using System.Net;
 global using System.Net.Http;
 global using System.Threading;
 global using System.Threading.Tasks;
@@ -10,3 +11,4 @@ global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Http;
 global using Microsoft.Extensions.Logging;
 global using Uno.Extensions.Http;
+global using Uno.Extensions.Logging;
