+++ src/Uno.Extensions.Http/SimpleAuthenticationToken.cs
@@ -1,8 +1,13 @@
 ﻿namespace Uno.Extensions.Http;
 
+/// <summary>
+/// Implements a simple authentication token provider
+/// </summary>
 public class SimpleAuthenticationToken : IAuthenticationTokenProvider
 {
+	/// <inheritdoc />
 	public string? AccessToken { get; set; }
 
+	/// <inheritdoc />
 	public Task<string> GetAccessToken() => Task.FromResult(AccessToken ?? string.Empty);
 }
